man: Provider manpages included in manpage generation
Previously manpages of providers were not compiled on

```
`make nroff`
```

This was because the manpages for the providers were not included in the
list of manpages to compile `real_man_pages`.
Provider manpages were included in `prov_install_man_pages`.
In the `Makefile.am` the `prov_install_man_pages` is an empty list.
However after configuring, the list is populated in the Makefile.

Signed-off-by: Chang Hyun Park <heartinpiece@gmail.com>